+++ js/dom-nodes.js
@@ -1,0 +1,14 @@
+export const range = document.querySelector('#range');
+export const form = document.querySelector('.ad-form');
+export const roomsField = form.querySelector('[name="rooms"]');
+export const capacityField = form.querySelector('[name="capacity"]');
+export const priceField = document.querySelector('#price');
+export const formInputs = document.querySelectorAll('.ad-form__watch-focus');
+export const timeFieldset = document.querySelector('#time-fieldset');
+export const timeFields = timeFieldset.querySelectorAll('select');
+export const typeField = document.querySelector('#type');
+export const mapFilter = document.querySelector('.map__filters');
+export const filterRoomsField = document.querySelector('#housing-rooms');
+export const filterTypeField = document.querySelector('#housing-type');
+export const filterPriceField = document.querySelector('#housing-price');
+export const filterGuestsField = document.querySelector('#housing-guests');
