+++ js/main.js
@@ -3,8 +3,9 @@ import {mapLoaded, mainPin} from './libs/leaflet-init.js';
 import './validate-form.js';
 import {interactionWithForm} from './form.js';
 import {getData} from './api.js';
-import {placeAddsLabels, onMainPinMoveend} from './map.js';
+import {onMainPinMoveend, placeAddsLabels} from './map.js';
 import {showDataError} from './show-error-or-success.js';
+import {GETTING_DATA_URL} from './const.js';
 
 // Деактивировали интерфейс
 disableInterface();
@@ -15,7 +16,7 @@ if (mapLoaded) {
 }
 
 // Получаем данные с сервера, отрисовываем метки и обрабатываем ошибку загрузки
-getData(placeAddsLabels, showDataError);
+getData(GETTING_DATA_URL, placeAddsLabels, showDataError);
 
 // Взаимодействуем с главным пином
 mainPin.on('moveend', onMainPinMoveend);
