--- js/preset-const.js
@@ -1,32 +1,0 @@
-export const SIMILAR_ADD_COUNT = 10;
-export const COORDINATE_FRACTIONAL_DIGITS = 5;
-export const COUNT_FOR_CORRECT_END = 4;
-
-export const BASE_COORDS = {
-  lat: 35.67500,
-  lng: 139.75000
-};
-
-export const TYPES = {
-  flat: 'Квартира',
-  bungalow: 'Бунгало',
-  house: 'Дом',
-  palace: 'Дворец',
-  hotel: 'Отель'
-};
-
-export const TYPES_NIN_PRICE = {
-  bungalow: '0',
-  flat: '1000',
-  hotel: '3000',
-  house: '5000',
-  palace: '10000'
-};
-
-export const ZOOM = 12;
-
-export const FORM_ERROR_MESSAGE = 'Не удалось отправить форму. Попробуйте ещё раз';
-export const ROOMS_GUESTS_ERROR_MESSAGE = 'Выберите другое значение';
-
-export const GETTING_DATA_URL = 'https://25.javascript.pages.academy/keksobooking/data';
-export const SENDING_DATA_URL = 'https://25.javascript.pages.academy/keksobooking';
