+++ js/validate-form.js
@@ -1,7 +1,21 @@
 import './libs/noUiSlider-init.js';
 import {pristine} from './libs/pristin-init.js';
-import {range, capacityField, roomsField, roomsOption, guestsOption, valuesDivs, formInputs} from './const.js';
-import {ROOMS_GUESTS_ERROR_MESSAGE} from './preset-const.js';
+import {range, capacityField, roomsField, formInputs, priceField} from './dom-nodes.js';
+import {ROOMS_GUESTS_ERROR_MESSAGE} from './const.js';
+
+export const roomsOption = {
+  '1': ['1'],
+  '2': ['1', '2'],
+  '3': ['1', '2', '3'],
+  '100': ['0']
+};
+
+export const guestsOption = {
+  '1': ['1', '2', '3'],
+  '2': ['2', '3'],
+  '3': ['3'],
+  '0': ['100']
+};
 
 // Валидируем число комнат и гостей
 const validateRooms = () => roomsOption[roomsField.value].includes(capacityField.value);
@@ -12,22 +26,15 @@ pristine.addValidator(capacityField, validateGuests, ROOMS_GUESTS_ERROR_MESSAGE)
 
 // Привязываем значения ползунка к инпуту и убираем сообщение об ошибке, когда оно не нужно
 range.noUiSlider.on('update', (values, handle) => {
-  const roller = range.querySelector('.noUi-touch-area');
-  const rangeSliderElements = [roller, range];
+  priceField.value = values[handle];
 
-  rangeSliderElements.forEach((rangeSliderElement) => {
-    rangeSliderElement.addEventListener('click', () => {
-      valuesDivs[handle].value = values[handle];
-    });
-  });
-
-  if (isFinite(valuesDivs[handle].value) ) {
+  if (isFinite(priceField.value) ) {
     pristine.reset();
   }
 });
 
 // Функция для удаления ошибок пристины
-export const deleteErrors = (evt) => {
+export const onInputBlur = (evt) => {
 
   const errorMessage = evt.target.nextElementSibling;
 
@@ -40,5 +47,5 @@ export const deleteErrors = (evt) => {
 
 // Убираем сообщение об ошибке при отсутствии фокуса
 formInputs.forEach((input) => {
-  input.addEventListener('blur', deleteErrors);
+  input.addEventListener('blur', onInputBlur);
 });
